purge: added index compaction after the index is loaded (#743)
* purge: added index compaction after the index is loaded


* pebble: fixed edge case with invalid ranges

+++ pkg/core/purge_pebble.go
@@ -32,7 +32,16 @@ func (kv *kvPebble) Drop() error {
 	}()
 
 	start, end := iterator.RangeBounds()
+	if pebble.DefaultComparer.Compare(start, end) >= 0 {
+		return nil
+	}
+
 	if err := kv.DB.DeleteRange(start, end, &pebble.WriteOptions{Sync: false}); err != nil {
+		return err
+	}
+
+	// as DeleteRange excludes the upper bound
+	if err := kv.DB.Delete(end, &pebble.WriteOptions{Sync: false}); err != nil && !errors.Is(err, pebble.ErrNotFound) {
 		return err
 	}
 
@@ -88,6 +97,20 @@ func (kv *kvPebble) Set(key, value []byte) error {
 
 func (kv *kvPebble) SetIfNotExists(key, value []byte) error {
 	return kv.Merge(key, value, &pebble.WriteOptions{Sync: false}) // skip new value
+}
+
+func (kv *kvPebble) Compact() error {
+	iterator := kv.DB.NewIter(nil)
+	start, end := iterator.RangeBounds()
+	defer func() {
+		_ = iterator.Close()
+	}()
+
+	if pebble.DefaultComparer.Compare(start, end) >= 0 {
+		return nil
+	}
+
+	return kv.DB.Compact(start, end, true)
 }
 
 func (i *kvPebbleIterator) Next() bool {
